+++ application/forms/Command/Object/DeleteDowntimeForm.php
@@ -40,6 +40,10 @@ class DeleteDowntimeForm extends CommandForm
 
     protected function getCommand(Model $object)
     {
+        if (! $this->isGrantedOn('icingadb/command/downtime/delete', $object->{$object->object_type})) {
+            return null;
+        }
+
         $command = new DeleteDowntimeCommand();
         $command->setDowntimeName($object->name);
         $command->setAuthor($this->getAuth()->getUser()->getUsername());
